requirements.txt
streamlit
google-generativeai
pandas
gnews
newspaper3k
lxml_html_clean